Move waiters and workers into separate folders.
