+++ src/web/waiters/BackgroundWorkerWaiter.mjs
@@ -4,7 +4,7 @@
  * @license Apache-2.0
  */
 
-import ChefWorker from "worker-loader?inline&fallback=false!../core/ChefWorker";
+import ChefWorker from "worker-loader?inline&fallback=false!../../core/ChefWorker";
 
 /**
  * Waiter to handle conversations with a ChefWorker in the background.
