+++ src/web/waiters/HighlighterWaiter.mjs
