+++ src/web/waiters/OperationsWaiter.mjs
@@ -4,7 +4,7 @@
  * @license Apache-2.0
  */
 
-import HTMLOperation from "./HTMLOperation";
+import HTMLOperation from "../HTMLOperation";
 import Sortable from "sortablejs";
