+++ src/web/waiters/WorkerWaiter.mjs
@@ -5,7 +5,7 @@
  * @license Apache-2.0
  */
 
-import ChefWorker from "worker-loader?inline&fallback=false!../core/ChefWorker";
+import ChefWorker from "worker-loader?inline&fallback=false!../../core/ChefWorker";
 
 /**
  * Waiter to handle conversations with the ChefWorker
