+++ src/web/workers/LoaderWorker.js
